requirements.txt
faster-whisper
pyannote.audio>=3.1
torch
torchaudio
sounddevice
scipy
numpy
huggingface_hub
python-dotenv
